ML/requirements.txt
absl-py==0.7.1
astor==0.8.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.16.4
Pillow==6.1.0
protobuf==3.8.0
six==1.12.0
tb-nightly==1.14.0a20190603
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019060501
Werkzeug==0.15.4
wrapt==1.11.2

tensorflow-gpu==1.12.0
#tensorflow==1.14.0